+++ src/components/misc/infiniteScroll/InfiniteScroll.svelte
@@ -9,24 +9,33 @@
   import { onMount } from "svelte";
   import { observeSrolledToBottom, loadMore } from "./utils/initInfiniteSroll";
   import Button from "../../buttons/button/Button.svelte";
+  import Spinner from "../spinner/Spinner.svelte";
 
   let showButton = false;
+  let showSpinner = false;
+  let countPages = 1;
 
   const loadedMoreEvent = new CustomEvent("khao-infinite-scroll-loaded-more", {
     bubbles: true,
   });
 
   const handleScolledToBottom = () => {
-    showButton = true;
+    if (countPages < parseInt(numberOfPages, 10)) {
+      showButton = true;
+    }
   };
 
   const handleButtonClick = () => {
     loadMore(querySelector, paginationSlug, handleLoadedContent);
     showButton = false;
+    showSpinner = true;
+
+    countPages++;
   };
 
   const handleLoadedContent = () => {
     showButton = false;
+    showSpinner = false;
     window.dispatchEvent(loadedMoreEvent);
   };
 
@@ -39,18 +48,23 @@
   export let querySelector: string;
   export let paginationSlug: string;
   export let buttonLabel: string;
+  export let numberOfPages: string = "1";
 </script>
 
 <div bind:this={scroller} class="infinite-scroll">
   {#if showButton}
     <Button label={buttonLabel} onClick={handleButtonClick}></Button>
   {/if}
+  {#if showSpinner}
+    <Spinner />
+  {/if}
 </div>
 
 <style>
   .infinite-scroll {
+    display: flex;
+    justify-content: space-around;
     width: 100%;
-    display: block;
-    height: 2ren;
+    height: 2rem;
   }
 </style>